Added ContextAware Mixin class and simplified meta classes
